Add code for MT objective

+++ python/onmt/translate/beam_search.py
@@ -205,7 +205,7 @@ class BeamSearch(DecodeStrategy):
         # on real data (newstest2017) with the pretrained transformer,
         # it's faster to not move this back to the original device
         self.is_finished = self.is_finished.to('cpu')
-        self.top_beam_finished |= self.is_finished[:, 0].eq(1)
+        self.top_beam_finished |= self.is_finished[:, 0].eq(1).to(torch.uint8)
         predictions = self.alive_seq.view(_B_old, self.beam_size, step)
         attention = (
             self.alive_attn.view(
